implemented patrol and attack for ants

+++ backend/game_logic/ant.py
@@ -18,7 +18,7 @@ class Ant(BaseModel):
     age: int  # Age of the ant
     type: str  # Type of the ant (e.g., worker, soldier, queen)
     task: TaskEnum  # Current task of the ant
-    hp: int
+    hp: float
     coords: dict  # Absolute coordinates of the ant (e.g., {"x": -100, "y": 50})
     objective: str  # ID of the objective entity the ant is interacting with
     destination: str  # ID of the object the ant is heading to
