+++ backend/game_logic/enemy.py
@@ -13,5 +13,5 @@ class Enemy(BaseModel):
     id: str  # Unique identifier for the enemy
     type: EnemyTypeEnum  # Type of the enemy (e.g., maggot, mantis, beetle)
     coords: Dict[str, float]  # Coordinates of the enemy (e.g., {"x": 100, "y": 200})
-    hp: int  # Health points of the enemy
+    hp: float  # Health points of the enemy
     speed: float  # Speed of the enemy
